+++ src/pages/Background/index.ts
@@ -5,6 +5,13 @@ const main = async () => {
   console.log("kartrak - background script started");
   tabtimes();
   analyzes();
+
+  chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
+    if (request.action === "getTabId") {
+      const tabId = sender.tab?.id;
+      sendResponse({ tabId });
+    }
+  });
 };
 
 main();